Added Tinsel
also updated chart speed to be a float and multiplicative, in case we combine multiple chart speed values.

+++ Scenes/Puppets/Scripts/PlayerStats.cs
@@ -16,7 +16,7 @@ public partial class PlayerStats : Resource
     public int Rerolls = 0;
     public int Shortcuts = 0;
     public int DiscountPercent = 0;
-    public int ChartSpeedMultiplier = 1;
+    public float ChartSpeedMultiplier = 1f;
 
     //Array in order of descending rarities, Legendary -> ... Common. Int odds out of 100.
     public int[] RarityOdds = [1, 5, 10, 20, 100];
